refactor(color-select-panel): extract api init to index.ts
